requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.3
django-crispy-forms==1.11.2
django-import-export==2.5.0
django-xhtml2pdf==0.0.4
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
python-barcode==0.13.1
python-bidi==0.4.2
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.67
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
